requirements.txt
absl-py>=2.1.0
astunparse>=1.6.3
beautifulsoup4>=4.12.3
blinker>=1.9.0
certifi>=2024.12.14
charset-normalizer>=3.4.1
click>=8.1.8
deepface>=0.0.93
filelock>=3.17.0
fire>=0.7.0
Flask>=2.0.1
Flask-Cors>=5.0.0
flatbuffers>=25.1.21
gast>=0.6.0
gdown>=5.2.0
google-api-python-client>=2.38.0
google-generativeai>=0.1.0
google-pasta>=0.2.0
grpcio>=1.69.0
gunicorn>=23.0.0
h5py>=3.12.1
idna>=3.10
itsdangerous>=2.2.0
Jinja2>=3.1.5
joblib>=1.4.2
keras>=3.8.0
libclang>=18.1.1
lz4>=4.3.3
Markdown>=3.7
markdown-it-py>=3.0.0
MarkupSafe>=3.0.2
mdurl>=0.1.2
ml-dtypes>=0.4.1
mtcnn>=1.0.0
namex>=0.0.8
numpy>=1.23.0
opencv-python>=4.6.0
opt-einsum>=3.3.0
packaging>=24.0
pandas>=1.5.0
Pillow>=9.0.0
protobuf>=4.25.3
PySocks>=1.7.1
python-dateutil>=2.8.2
pytz>=2024.1
requests>=2.31.0
retina-face>=0.0.14
rich>=13.7.1
scipy>=1.10.1
setuptools>=68.2.2
six>=1.16.0
soupsieve>=2.5
tensorboard>=2.16.2
tensorflow>=2.12.0
termcolor>=2.4.0
tqdm>=4.66.4
typing_extensions>=4.12.2
tzdata>=2024.1
urllib3>=2.0.7
Werkzeug>=2.2.3
wheel>=0.43.0
wrapt>=1.17.2
matplotlib>=3.7.0
scikit-learn>=1.2.2
seaborn>=0.12.2
